--- src/app/app.firebase.config.ts
@@ -1,8 +1,0 @@
-const FIREBASE_CONFIG = {
-    apiKey: "AIzaSyBgB8GwKV_KG87G6FxmuhhYsJWiSMCB_6U",
-    authDomain: "fir-auth-3c10d.firebaseapp.com",
-    databaseURL: "https://fir-auth-3c10d.firebaseio.com",
-    projectId: "fir-auth-3c10d",
-    storageBucket: "fir-auth-3c10d.appspot.com",
-    messagingSenderId: "523649393350"
-  };
