+++ src/models/user.ts
@@ -1,4 +1,4 @@
-export interface user {
+export interface User {
   email : string;
   password : string;
 }
