+++ src/pages/login/login.html
@@ -26,7 +26,7 @@
       <ion-input type="password" [(ngModel)] = "user.password" ></ion-input>
     </ion-item>
     <div class="loginButton">
-      <button ion-button round outline (click) = "login()">Login</button>
+      <button ion-button round outline (click) = "login(user)">Login</button>
     </div>
