+++ src/pages/login/login.ts
@@ -1,6 +1,9 @@
 import { Component } from '@angular/core';
 import { IonicPage, NavController, NavParams } from 'ionic-angular';
-import {User} from './models/user';
+import { User } from '../../models/user';
+import { AngularFireAuth} from '@angular/fire/auth';
+import {TabsPage} from '../tabs/tabs';
+
 /**
  * Generated class for the LoginPage page.
  *
@@ -15,13 +18,27 @@ import {User} from './models/user';
 })
 export class LoginPage {
 
-  user = {} as USER;
+  user = {} as User;
 
-  constructor(public navCtrl: NavController, public navParams: NavParams) {
+  constructor(public navCtrl: NavController, public navParams: NavParams ,private aFauth: AngularFireAuth) {
   }
 
   ionViewDidLoad() {
     console.log('ionViewDidLoad LoginPage');
+  }
+  async login(user){
+
+    try{
+      const result = this.aFauth.auth.signInWithEmailAndPassword(user.email, user.password);
+      console.log(result);
+
+      if(result){
+        this.navCtrl.push(TabsPage);
+      }
+    }
+    catch(e){
+      console.error(e);
+    }
   }
 
 }
