+++ src/pages/signup/signup.html
@@ -20,21 +20,16 @@
 
     <ion-item>
       <ion-label floating>Email</ion-label>
-      <ion-input type="text" value="" ></ion-input>
+      <ion-input type="text" [(ngModel)] = "user.email"></ion-input>
     </ion-item>
 
     <ion-item>
       <ion-label floating>Password(enter at least 6 characters)</ion-label>
-      <ion-input type="password" value=""></ion-input>
-    </ion-item>
-
-    <ion-item>
-      <ion-label floating>Re-enter Password</ion-label>
-      <ion-input type="password" value=""></ion-input>
+      <ion-input type="password" [(ngModel)] = "user.password"></ion-input>
     </ion-item>
 
     <div class="signupButton" text-center>
-      <button ion-button round outline (click) = "signup()">Signup</button>
+      <button ion-button round outline (click) = "signup(user)">Signup</button>
     </div>
