+++ src/pages/signup/signup.ts
@@ -1,6 +1,8 @@
 import { Component } from '@angular/core';
 import { IonicPage, NavController, NavParams } from 'ionic-angular';
-import {TabsPage} from '../tabs/tabs';
+import { TabsPage } from '../tabs/tabs';
+import { AngularFireAuth} from '@angular/fire/auth';
+import { User } from '../../models/user';
 /**
  * Generated class for the SignupPage page.
  *
@@ -15,15 +17,25 @@ import {TabsPage} from '../tabs/tabs';
 })
 export class SignupPage {
 
-  constructor(public navCtrl: NavController, public navParams: NavParams) {
+user = {} as User;
+
+  constructor(public navCtrl: NavController, public navParams: NavParams, private aFauth: AngularFireAuth) {
   }
 
   ionViewDidLoad() {
     console.log('ionViewDidLoad SignupPage');
   }
-  signup(){
-    this.navCtrl.push(TabsPage);
+  async signup(user : User){
+    try{
+      const result = await this.aFauth.auth.createUserWithEmailAndPassword(user.email, user.password);
+      if(result){
+        console.log("Signed up");
+        this.navCtrl.push(TabsPage);
+      }
+    }
+    catch(e){
+      console.error(e);
+    }
   }
-
 
 }
