+++ src/pages/welcome/welcome.html
@@ -7,7 +7,7 @@
 <ion-content padding id="welcome" text-center>
   <img src="assets/imgs/elvey-extra-txt.png">
   <div id="welcome_text" class="after_logo">WELCOME</div>
-<button ion-button round outline  class="button" (click) = "login()">Login</button>
+<button ion-button round outline  class="button" (click) = "loginPush()">Login</button>
 <br>
-<button ion-button round outline  class="button2" (click) = "signup()">Signup</button>
+<button ion-button round outline  class="button2" (click) = "signupPush()">Signup</button>
 </ion-content>
